Don't complain if the working and output folders don't exist when trying to delete them.

+++ s3pi Doc/trunk/msbuild.bat
@@ -3,8 +3,8 @@ set CopyTo=.\Help
 set OutputPath=C:\Windows\Temp\Help
 set WorkingPath=C:\Windows\Temp\HelpWork
 
-rmdir /s/q %OutputPath%
-rmdir /s/q %WorkingPath%
+if exist "%OutputPath%\" rmdir /s/q %OutputPath%
+if exist "%WorkingPath%\" rmdir /s/q %WorkingPath%
 
 c:\Windows\Microsoft.NET\Framework\v3.5\msbuild s3pi.shfbproj "/p:OutputPath=%OutputPath:\=\\%" "/p:WorkingPath=%WorkingPath:\=\\%"
 if ERRORLEVEL 1 goto FAIL
